requirements.txt
pyyaml
numpy
pillow
opencv
scikit-image
pytorch
torchvision
pydicom
matplotlib
albumentations
wandb